requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
slowapi==0.1.9
requests==2.31.0
openai==1.3.0
scikit-learn==1.3.2
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# PDF processing (minimal dependencies)
PyPDF2==3.0.1
